requirements.txt
langchain
fastapi
uvicorn
streamlit
pinecone 
langchain-pinecone 
langchain-groq 
python-dotenv
langchain-huggingface
langchain-text-splitters
sentence-transformers
docling
pandas
Pillow
tabulate
torch
python-magic